feat: implement CanvasClient component with drag-and-drop functionality and initial block data retrieval
